modules/genrams/xilinx/*: use grounded srstreg
The attributes DO_REG and EN_SYN are always
complementary to each other. So, when one of
them is 1 (or true) the other is 0 (or false)
and vice-versa.

However, as per Xilinx documentation and my own
experience (errors on Xilinx bitgen), in order to
use the RSTREG signal, both of them need to be 1
(or true), which is never the case for the current
mapping.

A fix is to ground the "srstreg" signal, as we never
have the necessary combination of attributes to use it,
anyway.

Signed-off-by: Dimitris Lampridis <Dimitris.Lampridis@cern.ch>